Add files via upload
The primary objective of this project is to develop a robust and reliable model that can accurately predict Australian vehicle prices. By harnessing the power of machine learning and data analysis, our goal is to provide a tool that not only assists potential buyers and sellers in making informed decisions but also contributes valuable insights to the automotive industry.